--- pkg/services/ngalert/migration/models/state.go
@@ -1,7 +1,0 @@
-package models
-
-// OrgMigrationState contains information about the state of an org migration.
-type OrgMigrationState struct {
-	OrgID          int64    `json:"orgId"`
-	CreatedFolders []string `json:"createdFolders"`
-}
