+++ pkg/services/ngalert/notifier/file_store_test.go
@@ -7,12 +7,10 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-
-	"github.com/grafana/grafana/pkg/services/ngalert/tests/fakes"
 )
 
 func TestFileStore_FilepathFor_DirectoryNotExist(t *testing.T) {
-	store := fakes.NewFakeKVStore(t)
+	store := NewFakeKVStore(t)
 	workingDir := filepath.Join(t.TempDir(), "notexistdir")
 	fs := NewFileStore(1, store, workingDir)
 	filekey := "silences"
@@ -33,7 +31,7 @@ func TestFileStore_FilepathFor_DirectoryNotExist(t *testing.T) {
 	}
 }
 func TestFileStore_FilepathFor(t *testing.T) {
-	store := fakes.NewFakeKVStore(t)
+	store := NewFakeKVStore(t)
 	workingDir := t.TempDir()
 	fs := NewFileStore(1, store, workingDir)
 	filekey := "silences"
@@ -75,7 +73,7 @@ func TestFileStore_FilepathFor(t *testing.T) {
 }
 
 func TestFileStore_Persist(t *testing.T) {
-	store := fakes.NewFakeKVStore(t)
+	store := NewFakeKVStore(t)
 	state := &fakeState{data: "something to marshal"}
 	workingDir := t.TempDir()
 	fs := NewFileStore(1, store, workingDir)
@@ -84,9 +82,9 @@ func TestFileStore_Persist(t *testing.T) {
 	size, err := fs.Persist(context.Background(), filekey, state)
 	require.NoError(t, err)
 	require.Equal(t, int64(20), size)
-	store.Mtx.Lock()
-	require.Len(t, store.Store, 1)
-	store.Mtx.Unlock()
+	store.mtx.Lock()
+	require.Len(t, store.store, 1)
+	store.mtx.Unlock()
 	v, ok, err := store.Get(context.Background(), 1, KVNamespace, filekey)
 	require.NoError(t, err)
 	require.True(t, ok)
